ansilo-connectors/jdbc-base/src/java/src/test/java/com/ansilo/connectors/data/NullDataTypeTest.java
package com.ansilo.connectors.data;

import static org.mockito.Mockito.times;
import static org.mockito.Mockito.verify;
import java.nio.ByteBuffer;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

public class NullDataTypeTest extends DataTypeTest {
    private NullDataType dataType;

    @BeforeEach
    void setUp() {
        super.setUp();
        this.dataType = new NullDataType();
    }

    @Test
    void testWrite() throws Exception {
        this.dataType.writeToByteBuffer(this.mapping, this.byteBuffer, this.resultSet, 0);

        verify(this.byteBuffer, times(1)).put((byte) 0);
    }

    @Test
    void testBindParam() throws Exception {
        var buff = ByteBuffer.allocate(1);
        buff.put((byte) 0);
        buff.rewind();
        this.dataType.bindParam(this.mapping, this.preparedStatement, 1, buff);

        verify(this.mapping, times(1)).bindNull(this.preparedStatement, 1,
                this.dataType.getTypeId());
    }
}
